Calendarコンポーネントから曜日を表示するコンポーネントを抜き出した
